feat: new monster, quicksilver elemental
I considered simply using quicksilver oozes or quicksilver dragons in
Dis, as both are reasonably strong late-game threats. However both
dragons and oozes are vulnerable to some of the collateral damage done
to allies by hell minions. It seemed like a better idea to make a
monster in a genus that won't fall so easily to tormennts, since the new
Dis list already has its share of living creatures.

This monster is something like a cross between a quicksilver dragon and
a shard shrike, with batty movement and qs breath, though not as fast as
a shrike.

Not 100% happy with repeating the shard shrike gimmick, could instead
just make quicksilver oozes torment immune. (How does one torment a blob
of quicksilver anyway?).