feat: new monster, wendigo
Wendigo are a new Cocytus monster. They're somewhat flimsy and act as
threat multipliers, with stunning burst and a new spell seracfall.
Seracfall turns a simulacrum in view into a very high power (3d43)
iceblast. Cocytus will have plenty of simulacra around as part of the
overhaul, but many simulacra are slow. The wendigo "help" the simulacra
along in a couple of ways. The targeting also means that a player facing
a wendigo plus simulacra will have to manage LOF from several
directions as the simulacra advance.