feat: define a cross-hell spawn population
To unify the hell branches, this defines a population subtable that
produces "cross-hell" threats: a flat (falling off in ood range)
sprinkling of tormentors and hellions bringing suffering to the damned,
plus rare lichen making sundry unholy pacts.

Lore note: Lichen, having mastered life and death, are not minions of
hell or bound there. They're simply visiting. So we don't want to put a
ton in any hell, even Tartarus.