feat: adjust derived undead attack flavours
This commit allows spectrals and simulacra to constrict and reach
(losing the flavor brand on that attack) and all derived undead to
trample.

These "physical" attack flavours mess with player positioning in various
ways; not including them on the higher tier derived undead made it
challenging to use those derived undead as serious threats.