vaults: adjust non-end Dis vault monsters and cleanup
Adjust monster sets of Dis vaults. When possible I tried to stick with
or switch to 098. When vault storyline demanded eg. a humanoid or a boss
monster one was chosen from the new Dis cast appropriately.

Occasionally some vaults desired slightly more variety, and juggernauts
get sprinkled in there. I'd considered putting them on the Dis spawn
table, but a few vault cameos seems to be better for them.